course_mcp/2. MCP with API/2. MCP with SQL/requirements.txt
fastmcp==3.1.1 # MCP 서버
mcp==1.26.0 
python-dotenv==1.2.2 # 환경 변수 관리
asyncpg==0.31.0 # PostgreSQL 연결
